requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.5.2
pydantic-settings==2.1.0
celery==5.3.6
redis==5.0.1
aiogram==3.2.1
python-dotenv==1.0.0
alembic==1.12.1
email-validator==2.1.0.post1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiohttp==3.9.1
tenacity==8.2.3
psycopg2-binary==2.9.9